+++ models/Players.py
@@ -1,11 +1,11 @@
 from models.BaseModel import Base
-from sqlalchemy import Column, Integer, Boolean, Float, ARRAY, String, ForeignKey
+from sqlalchemy import Column, Integer, Boolean, Float, ARRAY, String, ForeignKey, BigInteger
 
 
 class Player(Base):
 
     name = Column(String, nullable = False)
-    team_id = Column(Integer, ForeignKey('team.id'))
-    position_id = Column(Integer, ForeignKey('position.id'))
+    team_id = Column(BigInteger, ForeignKey('team.id'))
+    position_id = Column(BigInteger, ForeignKey('position.id'))
     year = Column(Integer, nullable=False, default=2014)
